% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_ascp.R
\name{download_data_ascp}
\alias{download_data_ascp}
\title{Title}
\usage{
download_data_ascp(
  SRR,
  outpath,
  Private_key = "/wrk/jiangdingkun/.aspera/connect/etc/asperaweb_id_dsa.openssh"
)
}
\arguments{
\item{Private_key}{}
}
\description{
Title
}
