% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_in_DNAstringSet.R
\name{enrich_in_DNAstringSet}
\alias{enrich_in_DNAstringSet}
\title{Title}
\usage{
enrich_in_DNAstringSet(
  PWMatrix,
  DNAStringSet,
  shuffle_num = 10,
  pseudo = 20,
  p.cutoff = 1e-05
)
}
\arguments{
\item{PWMatrix}{PWMatrixlist}

\item{DNAStringSet}{DNAStringSet}

\item{shuffle_num}{num}

\item{pseudo}{num}

\item{p.cutoff}{num}
}
\description{
Title
}
