% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_to_GO.R
\name{peak_to_GO}
\alias{peak_to_GO}
\title{Title}
\usage{
peak_to_GO(peak, top_peak_num = 2000, tssregion = c(-1000, 1000), go_num = 30)
}
\arguments{
\item{peak}{narrowPeak Grange object}

\item{top_peak_num}{num}

\item{go_num}{How many are displayed}
}
\description{
Title
}
