% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_track.R
\name{plot_track}
\alias{plot_track}
\title{plot_track}
\usage{
plot_track(
  zoom_region,
  bwfile,
  y_max = 40,
  track_color = "#80B1D3",
  peak_color = "#FB8072",
  sitename = "TF",
  show_gene = TRUE,
  show_peak = NULL,
  genome = NULL,
  ylables = c("10", "40"),
  tick.pos = c(10, 40),
  label_margin = 0.08,
  show_tatget_gene = NULL
)
}
\arguments{
\item{show_peak}{Grange object}

\item{Grange}{object}
}
\description{
plot_track
}
