% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_gr_grangse.R
\name{remove_gr_grangse}
\alias{remove_gr_grangse}
\title{Title remove_gr_grangse}
\usage{
remove_gr_grangse(b, minoverlap = 4)
}
\arguments{
\item{b}{Grange object}

\item{minoverlap}{Overlap at least a few bases}
}
\value{
Granges
}
\description{
Title remove_gr_grangse
}
\examples{
remove_gr_grangse(GRANGE,minoverlap=4)
}
