% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_logo.R
\name{TF_logo}
\alias{TF_logo}
\title{Title}
\usage{
TF_logo(pwmlist, outfile, weigh = NULL)
}
\arguments{
\item{pwmlist}{a PWMatrixList}

\item{outfile}{a pdf}

\item{weigh}{plot weigh}
}
\value{
pdf file
}
\description{
Title
}
